eliza_plugins/mri_3d_plugin/requirements.txt
numpy
nibabel
scikit-image
scipy
opencv-python-headless
plotly
google-generativeai
